test: isolate prefs DB per endpoint test to stop diagnostics state bleed
The Server.Tests endpoint factories shared the default zeus-prefs.db, so
persisted board/variant/override state from one test bled into the next
under the (already serial) suite — causing order-dependent failures in the
hardware/DSP diagnostics tests (e.g. a Hermes-marked board reporting the
OrionMkII receiver-bandwidth answer; GetLiveDiagnostics constraints missing).

Add a shared IsolatedPrefsFactory base that gives each WebApplicationFactory
its own unique ZEUS_PREFS_PATH temp DB (restored + deleted on Dispose) and
re-base the endpoint factories on it. Test-only; no product code touched.
Christian's board-diagnostics API is byte-for-byte unchanged.
