+++ tests/Zeus.Server.Tests/DisplayIntelligenceEndpointTests.cs
@@ -121,14 +121,12 @@ public sealed class DisplayIntelligenceEndpointTests : IDisposable
         Assert.Equal(HttpStatusCode.BadRequest, resp.StatusCode);
     }
 
-    private sealed class Factory(string dbPath) : WebApplicationFactory<Program>
+    private sealed class Factory(string dbPath) : IsolatedPrefsFactory
     {
-        protected override void ConfigureWebHost(IWebHostBuilder builder)
+        protected override void ConfigureExtra(IWebHostBuilder builder)
         {
-            builder.UseEnvironment("Test");
             builder.ConfigureServices(services =>
             {
-                services.RemoveAll<IHostedService>();
                 services.RemoveAll<DisplayIntelligenceSettingsStore>();
                 services.AddSingleton(sp => new DisplayIntelligenceSettingsStore(
                     sp.GetRequiredService<ILogger<DisplayIntelligenceSettingsStore>>(),
