+++ tests/Zeus.Server.Tests/G2OptionsEndpointTests.cs
@@ -96,33 +96,7 @@ public sealed class G2OptionsEndpointTests
         Assert.Contains("rx1AttenuatorDb", body.RootElement.GetProperty("error").GetString());
     }
 
-    private sealed class Factory : WebApplicationFactory<Program>
+    private sealed class Factory : IsolatedPrefsFactory
     {
-        private readonly string _dbPath = Path.Combine(
-            Path.GetTempPath(),
-            $"zeus-prefs-g2-options-{Guid.NewGuid():N}.db");
-        private readonly string? _previousPrefsPath;
-
-        public Factory()
-        {
-            _previousPrefsPath = Environment.GetEnvironmentVariable("ZEUS_PREFS_PATH");
-            Environment.SetEnvironmentVariable("ZEUS_PREFS_PATH", _dbPath);
-        }
-
-        protected override void ConfigureWebHost(IWebHostBuilder builder)
-        {
-            builder.UseEnvironment("Test");
-            builder.ConfigureServices(services =>
-            {
-                services.RemoveAll<IHostedService>();
-            });
-        }
-
-        protected override void Dispose(bool disposing)
-        {
-            base.Dispose(disposing);
-            Environment.SetEnvironmentVariable("ZEUS_PREFS_PATH", _previousPrefsPath);
-            try { if (File.Exists(_dbPath)) File.Delete(_dbPath); } catch { }
-        }
     }
 }
