+++ tests/Zeus.Server.Tests/HardwareKeyingEndpointTests.cs
@@ -106,15 +106,7 @@ public sealed class HardwareKeyingEndpointTests
         Assert.DoesNotContain("planned:/api/tx/external-ptt", controls);
     }
 
-    private sealed class Factory : WebApplicationFactory<Program>
+    private sealed class Factory : IsolatedPrefsFactory
     {
-        protected override void ConfigureWebHost(IWebHostBuilder builder)
-        {
-            builder.UseEnvironment("Test");
-            builder.ConfigureServices(services =>
-            {
-                services.RemoveAll<IHostedService>();
-            });
-        }
     }
 }
