+++ tests/Zeus.Server.Tests/TxLevelingEndpointValidationTests.cs
@@ -118,16 +118,14 @@ public class TxLevelingEndpointValidationTests : IClassFixture<TxLevelingEndpoin
         Assert.Equal(before?.CompressorGainDb, after?.CompressorGainDb);
     }
 
-    public sealed class Factory : WebApplicationFactory<Program>
+    public sealed class Factory : IsolatedPrefsFactory
     {
         private readonly MicGainEndpointTests.StubEngine _engine = new();
 
-        protected override void ConfigureWebHost(IWebHostBuilder builder)
+        protected override void ConfigureExtra(IWebHostBuilder builder)
         {
-            builder.UseEnvironment("Test");
             builder.ConfigureServices(services =>
             {
-                services.RemoveAll<IHostedService>();
                 services.RemoveAll<DspPipelineService>();
                 services.AddSingleton<DspPipelineService>(sp =>
                     new TestPipeline(
