+++ tests/Zeus.Server.Tests/UserIoEndpointTests.cs
@@ -177,15 +177,7 @@ public sealed class UserIoEndpointTests
         Assert.Contains("digIn.txDisableMappingStatus", telemetry);
     }
 
-    private sealed class Factory : WebApplicationFactory<Program>
+    private sealed class Factory : IsolatedPrefsFactory
     {
-        protected override void ConfigureWebHost(IWebHostBuilder builder)
-        {
-            builder.UseEnvironment("Test");
-            builder.ConfigureServices(services =>
-            {
-                services.RemoveAll<IHostedService>();
-            });
-        }
     }
 }
